// src/app/interfaces/IProducto.ts
export class IProducto {
    idProducto?: number|any;
    nombre?: string;
    descripcion?: string;
    idCategoria?: number;
    precio?: number;
    rutaImagen?: string;
    nombreImagen?: string;
    activo?: number;
    fechaRegistro?: string; // Usaremos una cadena para representar la fecha
  }
  